Add doxygen validator
- Add doxygen config
- Add script that checks if there are any doxygen warnings
- Add doxygen auto generated data to gitignore
- Add doxygen validator Github action
- Add or update all docblocks
- Update ubuntu to 22.04 in Github actions
- Install clang-tidy and clang-format 14
- Update clang tidy rule

+++ engine/src/liquid/renderer/MaterialPBR.h
@@ -5,8 +5,14 @@
 
 namespace liquid {
 
+/**
+ * @brief PBR material instance
+ */
 class MaterialPBR : public Material {
 public:
+  /**
+   * @brief PBR material properties
+   */
   struct Properties {
   public:
     /**
@@ -24,25 +30,84 @@ public:
     const std::vector<std::pair<String, Property>> getProperties() const;
 
   public:
+    /**
+     * Base color texture
+     */
     rhi::TextureHandle baseColorTexture = rhi::TextureHandle::Invalid;
-    int baseColorTextureCoord = -1;
+
+    /**
+     * Base color texture coordinate index
+     */
+    int8_t baseColorTextureCoord = -1;
+
+    /**
+     * Base color factor
+     */
     glm::vec4 baseColorFactor{};
 
+    /**
+     * Metallic roughness textures
+     */
     rhi::TextureHandle metallicRoughnessTexture = rhi::TextureHandle::Invalid;
-    int metallicRoughnessTextureCoord = -1;
+
+    /**
+     * Metallic roughness texture coordinate index
+     */
+    int8_t metallicRoughnessTextureCoord = -1;
+
+    /**
+     * Metallic factor
+     */
     float metallicFactor = 0.0f;
+
+    /**
+     * Roughness factor
+     */
     float roughnessFactor = 0.0f;
 
+    /**
+     * Normal texture
+     */
     rhi::TextureHandle normalTexture = rhi::TextureHandle::Invalid;
-    int normalTextureCoord = -1;
+
+    /**
+     * Normal texture coordinate index
+     */
+    int8_t normalTextureCoord = -1;
+
+    /**
+     * Normal scale
+     */
     float normalScale = 0.0f;
 
+    /**
+     * Occlusion texture
+     */
     rhi::TextureHandle occlusionTexture = rhi::TextureHandle::Invalid;
-    int occlusionTextureCoord = -1;
+
+    /**
+     * Occlusion texture coordinate index
+     */
+    int8_t occlusionTextureCoord = -1;
+
+    /**
+     * Occlusion strength
+     */
     float occlusionStrength = 0.0f;
 
+    /**
+     * Emissive texture
+     */
     rhi::TextureHandle emissiveTexture = rhi::TextureHandle::Invalid;
-    int emissiveTextureCoord = -1;
+
+    /**
+     * Emissive texture coordinate index
+     */
+    int8_t emissiveTextureCoord = -1;
+
+    /**
+     * Emissive factor
+     */
     glm::vec3 emissiveFactor{};
   };
